got rid of old testrun.py
